+++ packages/experiments/src/components/fluent/FluentStyles.ts
@@ -1,4 +1,10 @@
 import { FontWeights } from 'office-ui-fabric-react/lib/Styling';
+import { NeutralColors } from './FluentColors';
+import {
+  IChoiceGroupOptionStyleProps,
+  IChoiceGroupOptionStyles
+} from 'office-ui-fabric-react/lib/components/ChoiceGroup/ChoiceGroupOption';
+import { FontSizes } from './FluentType';
 
 /** Definitions for Depth, or shadow, levels. */
 const FluentDepthLevels = {
@@ -33,15 +39,9 @@ const FluentDepthLevels = {
   Level4: '0 16px 18px -4px rgba(0, 0, 0, 0.1)'
 };
 
-// const LinkStyles = {
-//   root: {
-//     // Styles
-//   }
-// };
-
 const BreadcrumbStyles = {
   itemLink: {
-    fontSize: '18px',
+    fontSize: FontSizes.size18,
     fontWeight: 400,
     selectors: {
       '&:last-child': {
@@ -53,23 +53,60 @@ const BreadcrumbStyles = {
 
 const PrimaryButtonStyles = {
   root: {
-    borderRadius: '2px',
-    boxShadow: FluentDepthLevels.Level1
+    borderRadius: '2px'
+    // boxShadow: FluentDepthLevels.Level1
   }
 };
 
 const CompoundButtonStyles = {
   root: {
-    borderRadius: '2px',
-    boxShadow: FluentDepthLevels.Level1
+    borderRadius: '2px'
+    // boxShadow: FluentDepthLevels.Level1
   }
 };
 
 const DefaultButtonStyles = {
   root: {
     borderRadius: '2px',
-    boxShadow: FluentDepthLevels.Level1
+    backgroundColor: '#fff',
+    border: `1px solid ${NeutralColors.gray20}`
+  },
+  rootHovered: {
+    backgroundColor: '#f3f2f1'
   }
+};
+
+const CheckboxStyles = {
+  checkbox: {
+    borderRadius: '2px'
+  }
+};
+
+const ChoiceGroupOptionStyles = (props: IChoiceGroupOptionStyleProps): IChoiceGroupOptionStyles => {
+  const { checked, hasIcon, hasImage } = props;
+  const radioButtonSpacing = 1;
+  const radioButtonInnerSize = 6;
+  return {
+    field: {
+      selectors: {
+        ':after': [
+          {
+            top: 4,
+            left: 4,
+            width: 12,
+            height: 12,
+            borderWidth: 6
+          },
+          checked &&
+            (hasIcon || hasImage) && {
+              top: radioButtonSpacing + radioButtonInnerSize,
+              right: radioButtonSpacing + radioButtonInnerSize,
+              left: 'auto' // To reset the value of 'left' to its default value, so that 'right' works
+            }
+        ]
+      }
+    }
+  };
 };
 
 const DialogStyles = {
@@ -84,14 +121,17 @@ const DialogContentStyles = {
   }
 };
 
+const LabelStyles = {
+  root: {
+    fontWeight: FontWeights.semibold
+  }
+};
+
 // Roll up all style overrides in a single "Fluent theme" object
 export const FluentStyles = {
   Breadcrumb: {
     styles: BreadcrumbStyles
   },
-  // Link: {
-  //   styles: LinkStyles
-  // },
   PrimaryButton: {
     styles: PrimaryButtonStyles
   },
@@ -101,33 +141,19 @@ export const FluentStyles = {
   CompoundButton: {
     styles: CompoundButtonStyles
   },
+  Checkbox: {
+    styles: CheckboxStyles
+  },
+  ChoiceGroupOption: {
+    styles: ChoiceGroupOptionStyles
+  },
   Dialog: {
     styles: DialogStyles
   },
   DialogContent: {
     styles: DialogContentStyles
+  },
+  Label: {
+    styles: LabelStyles
   }
 };
-
-// export class FluentStylesBasicExample extends React.Component<{}, {}> {
-//   public render(): JSX.Element {
-//     return (
-//       <div>
-//         <h2>Link</h2>
-//         <h3>Current theme</h3>
-//         <FluentThemeLinkExample />
-//         <h3>Fluent theme</h3>
-//         <Customizer scopedSettings={{ ...FluentStyles }}>
-//           <FluentThemeLinkExample />
-//         </Customizer>
-//         <h2>Buttons</h2>
-//         <h3>Current theme</h3>
-//         <FluentThemeButtonExample />
-//         <h3>Fluent theme</h3>
-//         <Customizer scopedSettings={{ ...FluentStyles }}>
-//           <FluentThemeButtonExample />
-//         </Customizer>
-//       </div>
-//     );
-//   }
-// }