+++ packages/office-ui-fabric-react/src/components/Callout/Callout.types.ts
@@ -3,7 +3,7 @@ import { DirectionalHint } from '../../common/DirectionalHint';
 import { IPoint, IRectangle, IStyleFunctionOrObject } from '../../Utilities';
 import { ICalloutPositionedInfo } from '../../utilities/positioning';
 
-export interface ICallout { }
+export interface ICallout {}
 
 export interface ICalloutProps {
   /**