+++ packages/office-ui-fabric-react/src/components/MessageBar/MessageBar.types.ts
@@ -1,7 +1,9 @@
 import * as React from 'react';
 import { BaseButton, Button } from '../../Button';
+import { ITheme, IStyle } from '../../Styling';
+import { IStyleFunctionOrObject } from '../../Utilities';
 
-export interface IMessageBar {}
+export interface IMessageBar { }
 
 export interface IMessageBarProps extends React.HTMLAttributes<HTMLElement> {
   /**
@@ -59,6 +61,120 @@ export interface IMessageBarProps extends React.HTMLAttributes<HTMLElement> {
    * Aria label on overflow button if truncated is defined.
    */
   overflowButtonAriaLabel?: string;
+
+  /**
+   * Additional CSS class(es) to apply to the MessageBar.
+   */
+  className?: string;
+
+  /**
+   * Theme (provided through customization.)
+   */
+  theme?: ITheme;
+
+  /**
+   * Call to provide customized styling that will layer on top of the variant rules.
+   */
+  styles?: IStyleFunctionOrObject<IMessageBarStyleProps, IMessageBarStyles>;
+}
+
+export interface IMessageBarStyleProps {
+  /**
+   * Theme (provided through customization).
+   */
+  theme: ITheme;
+
+  /**
+   * Additional CSS class(es).
+   */
+  className?: string;
+
+  /**
+   * Type of the MessageBar.
+   */
+  messageBarType?: MessageBarType;
+
+  /**
+   * Whether the MessageBar contains a dismiss button.
+   */
+  onDismiss?: boolean;
+
+  /**
+   * Whether the text is truncated.
+   */
+  truncated?: boolean;
+
+  /**
+   * Whether the MessageBar is rendered in multi line (as opposed to single line) mode.
+   */
+  isMultiline?: boolean;
+
+  /**
+   * Whether the single line MessageBar is being expanded.
+   */
+  expandSingleLine?: boolean;
+
+  /**
+   * Whether the MessageBar contains any action elements.
+   */
+  actions?: boolean;
+}
+
+export interface IMessageBarStyles {
+  /**
+   * Style set for the root element.
+   */
+  root?: IStyle;
+
+  /**
+   * Style set for the element containing the icon, text, and optional dismiss button.
+   */
+  content?: IStyle;
+
+  /**
+   * Style set for the element containing the icon.
+   */
+  iconContainer?: IStyle;
+
+  /**
+   * Style set for the icon.
+   */
+  icon?: IStyle;
+
+  /**
+   * Style set for the element containing the text.
+   */
+  text?: IStyle;
+
+  /**
+   * Style set for the text.
+   */
+  innerText?: IStyle;
+
+  /**
+   * Style set for the optional dismiss button.
+   */
+  dismissal?: IStyle;
+
+  /**
+   * Style set for the icon used to expand and collapse the MessageBar.
+   */
+  expand?: IStyle;
+
+  /**
+   * Style set for the element containing the dismiss button.
+   */
+  dismissSingleLine?: IStyle;
+
+  /**
+   * Style set for the element containing the expand icon.
+   */
+  expandSingleLine?: IStyle;
+
+  /**
+   * Style set for the optional element containing the action elements.
+   */
+  actions?: IStyle;
 }
 
 export enum MessageBarType {